+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## v5.50.0 - 2023-03-01
+
+### Changed
+
+- Cluster Security has increased health
+- Foundation planets feature more water
+
 ## v5.49.1 - 2023-02-23
 
 ### Fixed